Moves TrainRequestBody into 'sdk_messages.py'. Deletes 'training_request_body.py'
