+++ graphgrid_sdk/ggcore/sdk_messages.py
@@ -2,6 +2,7 @@
 
 import json
 import typing
+from dataclasses import dataclass
 
 import requests
 
@@ -299,3 +300,17 @@ class CheckTokenResponse(SdkServiceResponse):
 
     def __init__(self, generic_response: GenericResponse):
         super().__init__(generic_response)
+
+
+# pylint: disable=invalid-name
+@dataclass
+class TrainRequestBody:
+    """Store Airflow configuration json/request bodies"""
+    model: str
+    datasets: typing.Union[dict, str]
+    no_cache: bool = False
+    GPU: bool = False
+
+    def to_json(self):
+        """Encode TrainRequestBody to a json object"""
+        return json.dumps(self.__dict__, indent=4)
